chore: merged factories into one file
